Add additional mannual padding for CHERI.
The macro ANGLE_ENABLE_STRUCT_PADDING_WARNINGS enables -Wpadded. The
greater width and stronger alignment requirements of CHERI capabilities
results in more padding being added to structs than is currently
accounted for with the addition of manually added padding fields.

There are two options here. 1) Retain the warning but no longer generate
an error -Wno-error=padded or 2) Add further manual padding. For now I
have chosen to extend the manual padding. However, it should be noted
that this masks the impact that CHERI has on the size of these
structures and the implications on performance,